+++ services/web/server/src/simcore_service_webserver/security/_authorization.py
@@ -13,7 +13,7 @@ from servicelib.aiohttp.aiopg_utils import PostgresRetryPolicyUponOperation
 from simcore_postgres_database.models.users import UserRole
 from tenacity import retry
 
-from ..db.db_models import UserStatus, users
+from ..db.models import UserStatus, users
 from ..db.plugin import get_database_engine
 from ._access_model import ContextType, RoleBasedAccessModel, check_access
