+++ services/web/server/src/simcore_service_webserver/users/_db.py
@@ -5,7 +5,7 @@ from models_library.users import GroupID, UserID
 from simcore_postgres_database.models.users import UserStatus, users
 from sqlalchemy.sql import func
 
-from ..db.db_models import user_to_groups
+from ..db.models import user_to_groups
 
 
 async def do_update_expired_users(conn: SAConnection) -> list[UserID]:
