+++ services/web/server/tests/unit/with_dbs/02/test_projects_ports_handlers.py
@@ -18,7 +18,7 @@ from pytest_simcore.helpers.faker_webserver import (
 from pytest_simcore.helpers.utils_assert import assert_status
 from pytest_simcore.helpers.utils_login import UserInfoDict
 from simcore_service_webserver._meta import API_VTAG as VX
-from simcore_service_webserver.db.db_models import UserRole
+from simcore_service_webserver.db.models import UserRole
 from simcore_service_webserver.projects import _handlers_project_ports
 from simcore_service_webserver.projects.models import ProjectDict
 from yarl import URL
