+++ services/web/server/tests/unit/with_dbs/03/login/test_login_2fa.py
@@ -18,7 +18,7 @@ from pytest_simcore.helpers.utils_login import NewUser, parse_link, parse_test_m
 from servicelib.utils_secrets import generate_passcode
 from simcore_postgres_database.models.products import ProductLoginSettingsDict, products
 from simcore_service_webserver.application_settings import ApplicationSettings
-from simcore_service_webserver.db.db_models import UserStatus
+from simcore_service_webserver.db.models import UserStatus
 from simcore_service_webserver.login._2fa import (
     _do_create_2fa_code,
     create_2fa_code,
