+++ services/web/server/tests/unit/with_dbs/03/version_control/conftest.py
@@ -24,7 +24,7 @@ from simcore_postgres_database.models.projects_version_control import (
     projects_vc_snapshots,
 )
 from simcore_service_webserver._meta import API_VTAG as VX
-from simcore_service_webserver.db.db_models import UserRole
+from simcore_service_webserver.db.models import UserRole
 from simcore_service_webserver.db.plugin import APP_DB_ENGINE_KEY
 from simcore_service_webserver.log import setup_logging
 from simcore_service_webserver.projects.models import ProjectDict
